feat: Improve how events work
Added new methods for adding events and adding default events. Events
are now stored each as their own file in lib/events/, where they each
export an unbound, anonymous function for the action on the event. When
the events are added, they are put into a collection on the client for
later viewing and debugging.

undefined