refactor: Add args property to Command
Added an args property to Command class, which allows for a nicer syntax
for adding arguments to commands. Also added the execute callback to
Command, which processes args (if any), catches ArgumentErrors (if any),
throws other errors and runs the function. Updated default message event
to reflect these changes

undefined